Notification component update in order to support notification hub format replaces the hardcoded notifications
