+++ src/styles/notify/admin-bar.scss
@@ -22,7 +22,7 @@
     font-size: 0;
 
     /* WP Editor / Alert Red */
-    background: var(--wp-notify_color-red-alert);
+    background: var(--wp-notify_color-red);
 
     /* Classic / Dark Gray 800 */
     border: 1px solid var(--wp-notify_color-light-gray-800);
