+++ src/styles/notify/bullet.scss
@@ -21,8 +21,8 @@ $bullet-width: 6px;
 }
 
 #wpadminbar .wp-notification.unread >:first-child:before {
-  top: calc(50% - #{$bullet-width * .5 });
-  left: #{ (-24px - $bullet-width) * .5 }; // 9 - left padding - border /2  + bullet/2
+  top: calc( var(--wp-notify_hub-image-size) / 2  );
+  left: calc( var(--wp-notify_hub-image-size) * -1.6 - #{$bullet-width * .5 } );
 }
 #wpbody .wp-notification-wrap > :first-child:before {
   top: calc(10px - #{$bullet-width * .5 });
